fix: break cyclic import between hed_validator and hedtools_validator
Extract ValidationIssue and ValidationResult dataclasses into
validation_types.py to eliminate the circular dependency that CodeQL
flagged as py/unsafe-cyclic-import.
